Refactor adapter config modules
